Redirected triggerEval to camd-services
